backend/requirements.txt
Django>=4.2,<4.3
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<5.4

